Move the constant
